+++ src/main/java/umc/duckmelang/domain/member/controller/ProfileRestController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import umc.duckmelang.domain.member.facade.ProfileFacadeService;
 import umc.duckmelang.domain.member.dto.MemberResponseDto;
+import umc.duckmelang.domain.member.validation.annotation.ExistMember;
 import umc.duckmelang.domain.member.validation.annotation.ExistsMember;
 import umc.duckmelang.domain.memberprofileimage.converter.MemberProfileImageConverter;
 import umc.duckmelang.domain.memberprofileimage.domain.MemberProfileImage;
@@ -18,7 +19,14 @@ import umc.duckmelang.domain.post.converter.PostConverter;
 import umc.duckmelang.domain.post.domain.Post;
 import umc.duckmelang.domain.post.dto.PostResponseDto;
 import umc.duckmelang.domain.post.service.PostQueryService;
+import umc.duckmelang.domain.review.converter.ReviewConverter;
+import umc.duckmelang.domain.review.domain.Review;
+import umc.duckmelang.domain.review.dto.ReviewResponseDto;
+import umc.duckmelang.domain.review.service.ReviewQueryService;
+import umc.duckmelang.global.annotations.CommonApiResponses;
 import umc.duckmelang.global.apipayload.ApiResponse;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -27,6 +35,7 @@ public class ProfileRestController {
     private final ProfileFacadeService profileFacadeService;
     private final MemberProfileImageQueryService memberProfileImageQueryService;
     private final PostQueryService postQueryService;
+    private final ReviewQueryService reviewQueryService;
 
     @GetMapping(path = "/{memberId}")
     @Operation(summary = "다른 멤버 프로필 조회",description = "path variable로 프로필을 조회하고자하는 상대 member의 id를 받습니다.")
@@ -49,5 +58,13 @@ public class ProfileRestController {
         }
         Page<Post> postList = postQueryService.getPostListByMember(memberId, page);
         return ApiResponse.onSuccess(PostConverter.postPreviewListDto(postList));
+    }
+
+    @GetMapping("/{memberId}/reviews")
+    @CommonApiResponses
+    @Operation(summary = "다른 사람의 동행 후기 조회 API", description = "다른 사람의 프로필에서 동행 후기 볼 때 이용하는 API 입니다. 성별은 true일때 남자, false일때 여자입니다.")
+    public ApiResponse<ReviewResponseDto.ReviewListDto> getOtherReviewList(@ExistMember @PathVariable(name="memberId") Long memberId){
+        List<Review> reviewList = reviewQueryService.getReviewList(memberId);
+        return ApiResponse.onSuccess(ReviewConverter.reviewListDto(reviewList));
     }
 }
