+++ src/main/java/umc/duckmelang/domain/post/controller/PostController.java
@@ -35,7 +35,7 @@ public class PostController {
     private final PostCommandService postCommandService;
     private final BookmarkCommandService bookmarkCommandService;
 
-    @GetMapping("/")
+    @GetMapping("")
     @CommonApiResponses
     @Operation(summary = "홈화면 게시글 전체 조회 API", description = "조건 없이 모든 게시글을 조회하는 API 입니다. 페이징을 포함하며 한 페이지 당 10개 게시글을 보여줍니다. query String으로 page 번호를 주세요. page 번호는 0부터 시작합니다")
     public ApiResponse<PostResponseDto.PostPreviewListDto> getPostList (@ValidPageNumber @RequestParam(name = "page",  defaultValue = "0") Integer page){
@@ -81,13 +81,6 @@ public class PostController {
 
     }
 
-    @PostMapping("/{postId}/bookmarks")
-    @CommonApiResponses
-    @Operation(summary = "게시글 스크랩 API", description = "게시글 스크랩하는 API 입니다. 우선 memberId를 받습니다(추후 JWT로 변경 예정)")
-    public ApiResponse<BookmarkResponseDto.BookmarkJoinResultDto>joinBookmark (@PathVariable(name="postId") Long postId, @RequestParam(name="memberId") Long memberId){
-        Bookmark bookmark = bookmarkCommandService.joinBookmark(postId, memberId);
-        return ApiResponse.onSuccess(BookmarkConverter.bookmarkJoinResultDto(bookmark));
-    }
 
     @GetMapping("/my")
     @CommonApiResponses
