+++ src/main/java/umc/duckmelang/domain/review/controller/ReviewController.java
@@ -29,12 +29,13 @@ import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/reviews")
 @Validated
 public class ReviewController {
     private final ReviewCommandService reviewCommandService;
     private final ReviewQueryService reviewQueryService;
 
-    @PostMapping("/reviews")
+    @PostMapping("")
     @CommonApiResponses
     @Operation(summary = "후기글 작성 API", description = "후기글 작성 API입니다. Sender(memberId), Receiver(receiverId)를 써주세요. applicationId는 후기글 작성 페이지 내 관련 정보 조회 API에서 얻어오면 됩니다")
     public ApiResponse<ReviewResponseDto.ReviewJoinResultDto> joinReview (@RequestParam(name="memberId") Long memberId, @RequestBody @Valid ReviewRequestDto.ReviewJoinDto request){
@@ -42,23 +43,8 @@ public class ReviewController {
         return ApiResponse.onSuccess(ReviewConverter.reviewJoinResultDto(review));
     }
 
-    @GetMapping("/profile/reviews")
-    @CommonApiResponses
-    @Operation(summary = "나와의 동행 후기 조회 API", description = "내 프로필에서 나와의 동행 후기 볼 때 이용하는 API 입니다. 성별은 true일때 남자, false일때 여자입니다. memberId는 추후 JWT 변경 예정")
-    public ApiResponse<ReviewResponseDto.ReviewListDto> getMyReviewList(@RequestParam(name="memberId") Long memberId){
-        List<Review> reviewList = reviewQueryService.getReviewList(memberId);
-        return ApiResponse.onSuccess(ReviewConverter.reviewListDto(reviewList));
-    }
 
-    @GetMapping("/profile/{memberId}/reviews/")
-    @CommonApiResponses
-    @Operation(summary = "다른 사람의 동행 후기 조회 API", description = "다른 사람의 프로필에서 동행 후기 볼 때 이용하는 API 입니다. 성별은 true일때 남자, false일때 여자입니다.")
-    public ApiResponse<ReviewResponseDto.ReviewListDto> getOtherReviewList(@ExistMember @PathVariable(name="memberId") Long memberId){
-        List<Review> reviewList = reviewQueryService.getReviewList(memberId);
-        return ApiResponse.onSuccess(ReviewConverter.reviewListDto(reviewList));
-    }
-
-    @GetMapping("/reviews/information")
+    @GetMapping("/information")
     @CommonApiResponses
     @Operation(summary = "후기글 작성 페이지 내 관련 정보 조회 API", description = "후기글 작성 페이지에서 applicationId 외에 유저네임, 게시글 제목, 행사 날짜 등 정보를 보여주는 API 입니다. memberId를 requestParam으로 넣어주세요. myId는 추후 JWT 추출 예정")
     public ApiResponse<ReviewResponseDto.ReviewInformationDto> getReviewInformation(@ExistMember @RequestParam(name="memberId") Long memberId, @RequestParam(name="myId") Long myId){
